fix(tests): WSL2 support
This commit includes relevant changes in order to execute the
current test suite correctly on WSL2. This is done by checking for
an environment variable `WSL_DISTRO_NAME`. If it exists we're
for sure on WSL.

To include this in the tests a custom marker was introduced to
skip tests that are not running on WSL2 (as of 2021/01/26). The
reason is that there is no `systemd` support which `iptables` is using
in starting the service via `systemctl`.

To skip tests that are not runnable on WSL just mark them with
`@pytest.mark.skip_wsl`